Preliminary S-E-I-R Model
This is a S-E-I-R model that I built with Stella Architect software for my colleagues at the Ministry of Health.
Ibadan, Oyo State, Nigeria. Time horizon for the model is 1 year (365 days): 2020-03-01 through 2021-02-28. This scenario is based on an initial susceptible population of 7.5 million (estimated population of Oyo State), with Oyo State patient zero assumed to be exposed to COVID-19 on 2020-03-01 (Day 1). Scenarios show effects of a partial lockdown (night curfew) and boarder shutdown. Model structure and parameter values were adapted from https://covidactnow.org/. Major assumptions are that 4% of exposed population will develop mild to moderate symptoms of the disease and that, among these cases, 6% will perish. Please feel free to apply this model to your own local or regional planning efforts.